refactor: move landing page and component in to the folder. Update app.jsx to match with the updated code. Change /home to be /explore.
