--- client/src/App.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import CanvasAnimation from './components/CanvasAnimation';
-import Navigation from './components/Navigation';
-
-function App() {
-  return (
-    <>
-      <Navigation />
-      <CanvasAnimation />
-    </>
-  );
-}
-
-export default App;
