--- client/src/pages/HomePage.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function HomePage() {
-    return <div>HomePage</div>;
-}
-
-export default HomePage;
