--- client/src/pages/LandingPage.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function LandingPage() {
-    return <div>LandingPage</div>;
-}
-
-export default LandingPage;
