+++ client/src/pages/landingpage/LandingPage.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import CanvasAnimation from "../../components/CanvasAnimation";
+import Navigation from "../../components/Navigation";
+
+function LandingPage() {
+    return (
+        <>
+            <Navigation />
+            <CanvasAnimation />
+        </>
+    );
+}
+
+export default LandingPage;
